Added a contact form based on Form Add
Part of the code is explained on
http://parsimonycms.blogspot.fr/2013/01/creer-un-nouveau-bloc-exemple.html
[FRENCH]

This block is derived from Form Add.

It makes an customisable contact form predesigned.
*When the contact form is filled, it sends an mail ,to the mail
specified in the Configuration, with the fields filled by the user.
*I tested it with the main fields (wyiswyg,state,textarea,string,email).
*You can configured an recaptcha to prevent spam directly from the
Configuration interface.
*Notifications are predesigned and captcha error are notified.

Known Bugs:
--------------
If Recaptcha is active without a public/private key or a wrong
public/private key it will broke the admin page.
You can repair it by modifying the block at
/profiles/www/modules/core/views/NAMEOFTHEBLOCK.php

Possibles Improvements
----------------
*Send an confirmation to the user.
*Make the State field value, the actual value not the numeric one. (ex:
thisisthefirstvalue instead of 1) on the email
*Notification deactivation
*Notify multiple email.
*Translation
*Prettier notification email
*Last forms responded sended at a defined period instead of every
responded forms.
*Captcha Alternative?
*Captcha active after N tries.